requirements.txt
supabase
openai
python-dotenv
fastembed
pandas
pypdf
openpyxl
